requirements.txt
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
email-validator==2.2.0
flask==3.1.0
flask-login==0.6.3
flask-sqlalchemy==3.1.1
greenlet==3.2.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.6
markupsafe==3.0.2
packaging==25.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
requests==2.32.3
resend==2.9.0
sqlalchemy==2.0.40
typing-extensions==4.13.2
urllib3==2.4.0
werkzeug==3.1.3
